Separate distribute test
