+++ x/incentives/keeper/gauge_test.go
@@ -8,61 +8,6 @@ import (
 	lockuptypes "github.com/osmosis-labs/osmosis/x/lockup/types"
 )
 
-// TestDistribute tests that when the distribute command is executed on
-// a provided gauge,
-func (suite *KeeperTestSuite) TestDistribute() {
-	twoLockupUser := userLocks{
-		lockDurations: []time.Duration{defaultLockDuration, 2 * defaultLockDuration},
-		lockAmounts:   []sdk.Coins{defaultLPTokens, defaultLPTokens},
-	}
-	defaultGauge := perpGaugeDesc{
-		lockDenom:    defaultLPDenom,
-		lockDuration: defaultLockDuration,
-		rewardAmount: sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 3000)},
-	}
-	doubleLengthGauge := perpGaugeDesc{
-		lockDenom:    defaultLPDenom,
-		lockDuration: 2 * defaultLockDuration,
-		rewardAmount: sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 3000)},
-	}
-	oneKRewardCoins := sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 1000)}
-	twoKRewardCoins := sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 2000)}
-	fiveKRewardCoins := sdk.Coins{sdk.NewInt64Coin(defaultRewardDenom, 5000)}
-	tests := []struct {
-		users           []userLocks
-		gauges          []perpGaugeDesc
-		expectedRewards []sdk.Coins
-	}{
-		// gauge 1 gives 3k coins. Three locks, all eligible. 1k coins per lock
-		// so 1k to oneLockupUser, 2k to twoLockupUser
-		{
-			users:           []userLocks{oneLockupUser, twoLockupUser},
-			gauges:          []perpGaugeDesc{defaultGauge},
-			expectedRewards: []sdk.Coins{oneKRewardCoins, twoKRewardCoins},
-		},
-		// gauge 1 gives 3k coins. Three locks, all eligible.
-		// gauge 2 gives 3k coins to one lock, in twoLockupUser
-		// so 1k to oneLockupUser, 5k to twoLockupUser
-		{
-			users:           []userLocks{oneLockupUser, twoLockupUser},
-			gauges:          []perpGaugeDesc{defaultGauge, doubleLengthGauge},
-			expectedRewards: []sdk.Coins{oneKRewardCoins, fiveKRewardCoins},
-		},
-	}
-	for tcIndex, tc := range tests {
-		suite.SetupTest()
-		gauges := suite.SetupGauges(tc.gauges)
-		addrs := suite.SetupUserLocks(tc.users)
-		_, err := suite.app.IncentivesKeeper.Distribute(suite.ctx, gauges)
-		suite.Require().NoError(err)
-		// Check expected rewards
-		for i, addr := range addrs {
-			bal := suite.app.BankKeeper.GetAllBalances(suite.ctx, addr)
-			suite.Require().Equal(tc.expectedRewards[i].String(), bal.String(), "tcnum %d, person %d", tcIndex, i)
-		}
-	}
-}
-
 func (suite *KeeperTestSuite) TestInvalidDurationGaugeCreationValidation() {
 	suite.SetupTest()
 
@@ -78,85 +23,6 @@ func (suite *KeeperTestSuite) TestInvalidDurationGaugeCreationValidation() {
 	distrTo.Duration = defaultLockDuration
 	_, err = suite.app.IncentivesKeeper.CreateGauge(suite.ctx, false, addrs[0], defaultLiquidTokens, distrTo, time.Time{}, 1)
 	suite.Require().NoError(err)
-}
-
-// TODO: Make this test table driven, or move whatever it tests into
-// the much simpler TestDistribute
-func (suite *KeeperTestSuite) TestGetModuleToDistributeCoins() {
-	// test for module get gauges
-	suite.SetupTest()
-
-	// initial check
-	coins := suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, sdk.Coins(nil))
-
-	// setup lock and gauge
-	_, gaugeID, gaugeCoins, startTime := suite.SetupLockAndGauge(false)
-
-	// check after gauge creation
-	coins = suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, gaugeCoins)
-
-	// add to gauge and check
-	addCoins := sdk.Coins{sdk.NewInt64Coin("stake", 200)}
-	suite.AddToGauge(addCoins, gaugeID)
-	coins = suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, gaugeCoins.Add(addCoins...))
-
-	// check after creating another gauge from another address
-	_, _, gaugeCoins2, _ := suite.SetupNewGauge(false, sdk.Coins{sdk.NewInt64Coin("stake", 1000)})
-
-	coins = suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, gaugeCoins.Add(addCoins...).Add(gaugeCoins2...))
-
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// distribute coins to stakers
-	distrCoins, err := suite.app.IncentivesKeeper.Distribute(suite.ctx, []types.Gauge{*gauge})
-	suite.Require().NoError(err)
-	suite.Require().Equal(distrCoins, sdk.Coins{sdk.NewInt64Coin("stake", 105)})
-
-	// check gauge changes after distribution
-	coins = suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, gaugeCoins.Add(addCoins...).Add(gaugeCoins2...).Sub(distrCoins))
-}
-
-// TODO: Make this test table driven, or move whatever it tests into
-// the much simpler TestDistribute
-func (suite *KeeperTestSuite) TestGetModuleDistributedCoins() {
-	suite.SetupTest()
-
-	// initial check
-	coins := suite.app.IncentivesKeeper.GetModuleDistributedCoins(suite.ctx)
-	suite.Require().Equal(coins, sdk.Coins(nil))
-
-	// setup lock and gauge
-	_, gaugeID, _, startTime := suite.SetupLockAndGauge(false)
-
-	// check after gauge creation
-	coins = suite.app.IncentivesKeeper.GetModuleDistributedCoins(suite.ctx)
-	suite.Require().Equal(coins, sdk.Coins(nil))
-
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// distribute coins to stakers
-	distrCoins, err := suite.app.IncentivesKeeper.Distribute(suite.ctx, []types.Gauge{*gauge})
-	suite.Require().NoError(err)
-	suite.Require().Equal(distrCoins, sdk.Coins{sdk.NewInt64Coin("stake", 5)})
-
-	// check after distribution
-	coins = suite.app.IncentivesKeeper.GetModuleToDistributeCoins(suite.ctx)
-	suite.Require().Equal(coins, distrCoins)
 }
 
 // TODO: Make this test table driven
@@ -368,96 +234,6 @@ func (suite *KeeperTestSuite) TestPerpetualGaugeOperations() {
 	// check rewards estimation
 	rewardsEst = suite.app.IncentivesKeeper.GetRewardsEst(suite.ctx, lockOwner, []lockuptypes.PeriodLock{}, 100)
 	suite.Require().Equal(sdk.Coins(nil), rewardsEst)
-}
-
-func (suite *KeeperTestSuite) TestNoLockPerpetualGaugeDistribution() {
-	// test for module get gauges
-	suite.SetupTest()
-
-	// setup no lock perpetual gauge
-	coins := sdk.Coins{sdk.NewInt64Coin("stake", 10)}
-	gaugeID, _, _, startTime := suite.SetupNewGauge(true, coins)
-
-	// check gauges
-	gauges := suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
-	suite.Require().Len(gauges, 1)
-	expectedGauge := types.Gauge{
-		Id:          gaugeID,
-		IsPerpetual: true,
-		DistributeTo: lockuptypes.QueryCondition{
-			LockQueryType: lockuptypes.ByDuration,
-			Denom:         "lptoken",
-			Duration:      time.Second,
-		},
-		Coins:             coins,
-		NumEpochsPaidOver: 1,
-		FilledEpochs:      0,
-		DistributedCoins:  sdk.Coins{},
-		StartTime:         startTime,
-	}
-	suite.Require().Equal(gauges[0].String(), expectedGauge.String())
-
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// distribute coins to stakers, since it's perpetual distribute everything on single distribution
-	distrCoins, err := suite.app.IncentivesKeeper.Distribute(suite.ctx, []types.Gauge{*gauge})
-	suite.Require().NoError(err)
-	suite.Require().Equal(distrCoins, sdk.Coins(nil))
-
-	// check state is same after distribution
-	gauges = suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
-	suite.Require().Len(gauges, 1)
-	suite.Require().Equal(gauges[0].String(), expectedGauge.String())
-}
-
-func (suite *KeeperTestSuite) TestNoLockNonPerpetualGaugeDistribution() {
-	// test for module get gauges
-	suite.SetupTest()
-
-	// setup no lock non-perpetual gauge
-	coins := sdk.Coins{sdk.NewInt64Coin("stake", 10)}
-	gaugeID, _, _, startTime := suite.SetupNewGauge(false, coins)
-
-	// check gauges
-	gauges := suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
-	suite.Require().Len(gauges, 1)
-	expectedGauge := types.Gauge{
-		Id:          gaugeID,
-		IsPerpetual: false,
-		DistributeTo: lockuptypes.QueryCondition{
-			LockQueryType: lockuptypes.ByDuration,
-			Denom:         "lptoken",
-			Duration:      time.Second,
-		},
-		Coins:             coins,
-		NumEpochsPaidOver: 2,
-		FilledEpochs:      0,
-		DistributedCoins:  sdk.Coins{},
-		StartTime:         startTime,
-	}
-	suite.Require().Equal(gauges[0].String(), expectedGauge.String())
-
-	// start distribution
-	suite.ctx = suite.ctx.WithBlockTime(startTime)
-	gauge, err := suite.app.IncentivesKeeper.GetGaugeByID(suite.ctx, gaugeID)
-	suite.Require().NoError(err)
-	err = suite.app.IncentivesKeeper.BeginDistribution(suite.ctx, *gauge)
-	suite.Require().NoError(err)
-
-	// distribute coins to stakers, since it's perpetual distribute everything on single distribution
-	distrCoins, err := suite.app.IncentivesKeeper.Distribute(suite.ctx, []types.Gauge{*gauge})
-	suite.Require().NoError(err)
-	suite.Require().Equal(distrCoins, sdk.Coins(nil))
-
-	// check state is same after distribution
-	gauges = suite.app.IncentivesKeeper.GetNotFinishedGauges(suite.ctx)
-	suite.Require().Len(gauges, 1)
-	suite.Require().Equal(gauges[0].String(), expectedGauge.String())
 }
 
 func (suite *KeeperTestSuite) TestGaugesByDenom() {
